Tw/pbvc

+++ run_pipeline/run_pipeline.py
@@ -3,10 +3,10 @@ import os
 import shutil
 from pathlib import Path
 import multiprocessing
-from utils import getSessionID, getSubjectID, CopyandCheck, split_list
+from utils import getSessionID, getSubjectID, MoveandCheck, split_list
 from samseg_stats import generate_samseg_stats
 
-def process_samseg(dirs, derivatives_dir, freesurfer_path):
+def process_samseg(dirs, derivatives_dir, freesurfer_path, fsl_path, remove_temp=False):
 
     for dir in dirs:
 
@@ -27,7 +27,7 @@ def process_samseg(dirs, derivatives_dir, freesurfer_path):
         ### perform registartion with both T1w images
         # do the registration in a template folder and distribute its results to BIDS conform output directories later
         # create template folder
-        print(t1w)
+        #print(t1w)
         temp_dir = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w[0])}', 'temp')
         print(temp_dir)
         temp_dir_output = os.path.join(temp_dir, "output")
@@ -74,17 +74,31 @@ def process_samseg(dirs, derivatives_dir, freesurfer_path):
                     cd {temp_dir}; \
                     run_samseg_long {" ".join(map(str, cmd_arg))} --threads 4 --pallidum-separate --lesion --lesion-mask-pattern 0 1 -o output/\
                     ')
+        
+        ### run FSL-SIENA to calculate PBVC
+        os.system(f'FSLDIR={fsl_path};\
+                    . ${{FSLDIR}}/etc/fslconf/fsl.sh;\
+                    PATH=${{FSLDIR}}/bin:${{PATH}};\
+                    export FSLDIR PATH;\
+                    {fsl_path}/bin/siena {Path(t1w[0])} {Path(t1w[1])} -o {temp_dir} -B "-f 0.2 -B"')
 
         ### copy output files from temp folder to their session folders
         # write paths of output folders of the timepoint (tp) in a list
         tp_folder = sorted(list(str(x) for x in os.listdir(temp_dir_output) if "tp" in str(x)))
 
+        # copy the mean image file and pbvc files
+        mean_temp_location = os.path.join(temp_dir, "mean.mgz")
+        mean_target_location = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w_reg[0])}', f'sub-{getSubjectID(t1w_reg[0])}' + '_mean.mgz')
+        MoveandCheck(mean_temp_location, mean_target_location)
+        # copy the SIENA PBVC html report
+        pbvc_temp_location = os.path.join(temp_dir, "report.html")
+        pbvc_target_location = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w_reg[0])}', f'sub-{getSubjectID(t1w_reg[0])}' + '_PBVC-report.html')
+        MoveandCheck(pbvc_temp_location, pbvc_target_location)
+
         # only continue if more than one timepoint was segmented
+        # aggregate the samseg output files and move to appropriate directories
         if len(tp_folder) > 1:
-            # copy the mean image file
-            mean_temp_location = os.path.join(temp_dir, "mean.mgz")
-            mean_target_location = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w_reg[0])}', f'sub-{getSubjectID(t1w_reg[0])}' + '_mean.mgz')
-            CopyandCheck(mean_temp_location, mean_target_location)
+
             # iterate through all the timepoints 
             for i in range(len(tp_folder)):
                 # initialize empty lists
@@ -117,27 +131,27 @@ def process_samseg(dirs, derivatives_dir, freesurfer_path):
                 # copy files from template output folder to target output folder (one output folder per session)
                 # each time there is a check if the file in the temp folder exists and if it  was copied successfully
                 # T1w
-                CopyandCheck(t1w_reg_temp_location, t1w_reg_ses_location)
+                MoveandCheck(t1w_reg_temp_location, t1w_reg_ses_location)
                 # FLAIR
-                CopyandCheck(flair_reg_temp_location, flair_reg_ses_location)
+                MoveandCheck(flair_reg_temp_location, flair_reg_ses_location)
                 # FLAIR transformation file
-                CopyandCheck(flair_reg_field_temp_location, flair_reg_field_ses_location)
+                MoveandCheck(flair_reg_field_temp_location, flair_reg_field_ses_location)
                 # copy output files to target folder
                 for i in range(len(tp_files_temp_path)):
-                    CopyandCheck(tp_files_temp_path[i], tp_files_ses_path[i])
+                    MoveandCheck(tp_files_temp_path[i], tp_files_ses_path[i])
         else:
-            print(f'-- Only one or no timepoint in template folder, therefore no longitudinal data copied!')
+            print(f'Skipping longitudinal data copies.')
 
-        # delete the temp folder
-        shutil.rmtree(temp_dir)
-        if os.path.exists(temp_dir):
-            raise ValueError(f'failed to delete the template folder: {temp_dir}')
-        else:
-            print(f'successfully deleted the template folder: {temp_dir}')
-
+        if remove_temp:
+            # delete the temp folder
+            shutil.rmtree(temp_dir)
+            if os.path.exists(temp_dir):
+                raise ValueError(f'failed to delete the template folder: {temp_dir}')
+            else:
+                print(f'successfully deleted the template folder: {temp_dir}')
+ 
         # generate the actual samseg volumetric stats
-        # timepoint 1 segmentation
-        # deriv_ses = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w[i])}', f'ses-{getSessionID(t1w[i])}', 'anat')
+
         filename = f'sub-{getSubjectID(t1w[0])}_ses-{getSessionID(t1w[0])}_seg.mgz'
         bl_path = os.path.join(derivatives_dir, f'sub-{getSubjectID(t1w[0])}', f'ses-{getSessionID(t1w[0])}', 'anat', filename)
         filename = f'sub-{getSubjectID(t1w[1])}_ses-{getSessionID(t1w[1])}_seg.mgz'
@@ -150,10 +164,17 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Run SAMSEG Longitudinal Pipeline on cohort.')
     parser.add_argument('-i', '--input_directory', help='Folder of derivatives in BIDS database.', required=True)
     parser.add_argument('-n', '--number_of_workers', help='Number of parallel processing cores.', type=int, default=os.cpu_count()-1)
-    parser.add_argument('-f', '--freesurfer_path', help='Path to freesurfer binaries.', default='/home/twiltgen/Tun_software/Freesurfer/FS_7.3.2/freesurfer')
+    parser.add_argument('-f', '--freesurfer_path', help='Path to freesurfer binaries.', default='/home/jmcginnis/freesurfer')
+    parser.add_argument('-fsl', '--fsl_path', help='Path to FSL binaries.', default='/home/jmcginnis/fsl')
+    parser.add_argument('--remove_temp', action='store_true')
 
     # read the arguments
     args = parser.parse_args()
+
+    if args.remove_temp:
+        remove_temp = True
+    else:
+        remove_temp = False
 
     # generate derivatives/labels/
     derivatives_dir = os.path.join(args.input_directory, "derivatives/samseg-longitudinal-7.3.2")
@@ -169,7 +190,8 @@ if __name__ == "__main__":
     pool = multiprocessing.Pool(processes=args.number_of_workers)
     # creation, initialisation and launch of the different processes
     for x in range(0, args.number_of_workers):
-        pool.apply_async(process_samseg, args=(files[x], derivatives_dir, args.freesurfer_path))
+        pool.apply_async(process_samseg, args=(files[x], derivatives_dir, args.freesurfer_path, args.fsl_path, remove_temp))
 
     pool.close()
     pool.join()
+
